refactor workbench risk drawdown route
